+++ examples/arithmetic.hvmc
