+++ src/host.rs
@@ -11,7 +11,7 @@ use std::{
   ops::{Deref, DerefMut, RangeFrom},
 };
 
-pub mod encode;
+mod encode;
 
 /// Stores a bidirectional mapping between names and runtime defs.
 #[derive(Default)]
