+++ src/host/encode.rs
@@ -10,8 +10,8 @@ use crate::{
 };
 
 impl Host {
-  /// Encode a tree `tree` directly into a port or wire `trg`, skipping the
-  /// intermediate `Def` representation
+  /// Encode `tree` directly into `trg`, skipping the intermediate `Def`
+  /// representationw
   pub fn encode_tree<M: Mode>(&self, net: &mut run::Net<M>, trg: Trg, tree: &Tree) {
     EncodeState { host: self, net, vars: Default::default() }.encode(trg, tree);
   }
